fix(providers): Phase-1 money-correctness — Lithic Svix webhooks + Decimal card funding
Two money-path bugs from the provider audit:

1. Lithic webhook signature scheme. Lithic signs with Standard Webhooks
   (Svix), NOT a raw hex HMAC: webhook-id/webhook-timestamp/webhook-signature
   headers, whsec_ base64-decoded key, signed content "id.timestamp.body",
   base64 signature, space-delimited v1,<sig> entries, 5-min replay window.
   The treasury route was checking x-lithic-hmac as raw hex HMAC-SHA256 — it
   would reject every real Lithic webhook. Add LithicTreasuryClient.verify_webhook
   (Svix, fail-closed) and route through it; deprecate the raw-HMAC method
   (now always returns False to force migration). Regression test added.

2. Float money in the Lithic card-funding path. cards.fund summed USD balances
   as float(current) + float(payload.amount), corrupting precision on a money
   path. payload.amount is already Decimal; keep it Decimal through the sum and
   into the repo/adapter (signatures: float -> Decimal).

+++ apps/api/server/repositories/card_repository.py
@@ -12,6 +12,7 @@ from __future__ import annotations
 import os
 import uuid
 from datetime import UTC, datetime
+from decimal import Decimal
 from typing import Any
 
 
@@ -313,7 +314,7 @@ class CardRepository:
             return await self.get_by_card_id(card_id)
 
     async def update_funded_amount(
-        self, card_id: str, amount: float
+        self, card_id: str, amount: Decimal
     ) -> dict[str, Any] | None:
         if not self._use_postgres():
             card = self._cards.get(card_id)
